+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Administration/Unmute.cs
@@ -25,7 +25,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Administration
         [Remarks("<user>")]
         [RequireUserPermission(GuildPermission.ManageRoles)]
         [RequireUserPermission(GuildPermission.MuteMembers)]
-        public async Task UnmuteUser(IGuildUser user, [Remainder] string reason = null)
+        public async Task UnmuteUser(SocketGuildUser user, [Remainder] string reason = null)
         {
             Server server = await DatabaseQueries.GetOrCreateServerAsync(Context.Guild.Id);
             var mutedObject = await DatabaseQueries.GetFirstMatchAsync<MutedUser>(x => x.UserId == user.Id && x.ServerId == server.ServerId);
@@ -43,11 +43,13 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Administration
                 SocketRole muteRole = Context.Guild.Roles.FirstOrDefault(x => x.Name.ToLower() == "kaguya-mute");
                 await user.RemoveRoleAsync(muteRole);
 
+                // todo: Move away from using so many embeds. They take up space.
                 var embed = new KaguyaEmbedBuilder
                 {
                     Description = $"Successfully unmuted `{user}`"
                 };
 
+                KaguyaEvents.TriggerUnmute(new ModeratorEventArgs(server, Context.Guild, user, (SocketGuildUser)Context.User, reason));
                 await ReplyAsync(embed: embed.Build());
             }
             catch (NullReferenceException)